+++ internal/storage/storage.go
@@ -86,11 +86,6 @@ func (r *RoleCredentials) ExpireEpoch() int64 {
 	return time.UnixMilli(r.Expiration).Unix() // yes, millisec
 }
 
-// ExpireString returns the time the creds expire in the format of "2006-01-02 15:04:05.999999999 -0700 MST"
-func (r *RoleCredentials) ExpireString() string {
-	return time.UnixMilli(r.Expiration).String() // yes, millisec
-}
-
 // Expired returns if these role creds have expired or will expire in the next minute
 func (r *RoleCredentials) Expired() bool {
 	now := time.Now().Add(time.Minute).UnixMilli() // yes, millisec
@@ -98,7 +93,7 @@ func (r *RoleCredentials) Expired() bool {
 }
 
 // Return expire time in ISO8601 / RFC3339 format
-func (r *RoleCredentials) ExpireISO8601() string {
+func (r *RoleCredentials) ExpireString() string {
 	return time.Unix(r.ExpireEpoch(), 0).Format(time.RFC3339)
 }
 
@@ -106,7 +101,7 @@ func (r *RoleCredentials) ExpireISO8601() string {
 func (r *RoleCredentials) AccountIdStr() string {
 	s, err := utils.AccountIdToString(r.AccountId)
 	if err != nil {
-		log.WithError(err).Fatalf("Unable to parse accountId from AWS role credentials")
+		log.WithError(err).Fatalf("unable to parse accountId from AWS role credentials")
 	}
 	return s
 }
@@ -114,27 +109,27 @@ func (r *RoleCredentials) AccountIdStr() string {
 // Validate ensures we have the necessary fields
 func (r *RoleCredentials) Validate() error {
 	if r.RoleName == "" {
-		return fmt.Errorf("Missing roleName")
+		return fmt.Errorf("%s", "missing roleName")
 	}
 
 	if r.AccessKeyId == "" {
-		return fmt.Errorf("Missing accessKeyId")
+		return fmt.Errorf("%s", "missing accessKeyId")
 	}
 
 	if r.SecretAccessKey == "" {
-		return fmt.Errorf("Missing secretAccessKey")
+		return fmt.Errorf("%s", "missing secretAccessKey")
 	}
 
 	if r.AccountId == 0 {
-		return fmt.Errorf("Missing accountId")
+		return fmt.Errorf("%s", "missing accountId")
 	}
 
 	if r.SessionToken == "" {
-		return fmt.Errorf("Missing sessionToken")
+		return fmt.Errorf("%s", "missing sessionToken")
 	}
 
 	if r.Expiration == 0 {
-		return fmt.Errorf("Missing expiration")
+		return fmt.Errorf("%s", "missing expiration")
 	}
 	return nil
 }